Add github workflow for testing & linting all tools
Github provides us with 20 or 40 nodes that can run tests in parallel,
making it feasible to test the entire iuc repository.
planemo and Galaxy run on python 3.7 in this workflows, and tool
execution uses the `--biocontainers` flag which runs tools inside
containers where only the requirements listed in the tool xml is
available. All test results from individual chunks (tests are split
into 40 pieces) are combined at the end and can be downloaded at the
end of a run.

So let's try and fix all the currently broken tests (and missing tests)!

Once that is fone we can run weekly tests against
the latest galaxy stable branch and dev to identify changes in Galaxy
that break tool execution, and on the off-chance, dependencies that
have degraded over time (openssl not compatible anymore, resources that
went away, dependencies with "time bombs" ...).